backend/requirements.txt
Django~=4.2.0
djangorestframework~=3.15.1
django-cors-headers~=4.4.0
drf-yasg~=1.21.7
Pillow
gunicorn
psycopg2-binary

